refactor: modularize source structure and introduce route separation
- Restructured `lib/configressif` into modular directories:
  - `network/`, `ntp/`, `storage/`, `system/`, `utils/`, `http/`
  - Moved `accesspoint` logic into `network/accesspoint/`
  - Kept `configressif.{cpp,h}` at top-level as public entry point

- Introduced standalone route classes for API endpoints:
  - `NtpRoutes` handles `/api/ntp`
  - `NetworkRoutes` handles `/api/network`
  - Each route class accepts `AsyncWebServer&` via constructor

- Simplified `HttpServer` into a lightweight coordinator:
  - Manages only `begin()` and static file serving
  - Delegates all route binding to external route classes

- Improved encapsulation by removing inline lambda handlers from `httpserver.cpp`

- Prepared structure for future routes (system, auth) and middleware integration

Closes #11
