+++ workbench/lib/configressif/src/ntp/ntpstorage.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "storage.h"
+#include "storage/storage.h"
 #include <Preferences.h>
 
 namespace configressif
