+++ workbench/lib/configressif/src/utils/debugutils.cpp
@@ -1,4 +1,4 @@
-#include "debugutils.h"
+#include "utils/debugutils.h"
 #include <Arduino.h>
 
 void configressif::DebugUtils::countdown(unsigned int delaySeconds, const char *message)
